Add Event implementation for both go1.8 and go1.9 using syncmap
