+++ pushers/event/event.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"net"
 	"runtime/debug"
-	"sync"
-	"time"
 )
 
 //====================================================================================
@@ -77,62 +75,6 @@ type Option func(*Event)
 
 // Map defines a map type for event data.
 type Map map[string]interface{}
-
-// Event defines a object which adds key-value pairs into a map type for event data.
-type Event struct {
-	ml sync.Mutex
-	mp map[string]interface{}
-}
-
-// Add adds the key and value into the event.
-func (e *Event) Add(s string, v interface{}) {
-	e.ml.Lock()
-	e.mp[s] = v
-	e.ml.Unlock()
-}
-
-// Map returns the underline map for the giving object.
-func (e *Event) Map() Map {
-	var em Map
-
-	e.ml.Lock()
-	{
-		em = e.mp
-		e.mp = make(map[string]interface{})
-	}
-	e.ml.Unlock()
-
-	return em
-}
-
-// Get retrieves a giving value for a key has string.
-func (e *Event) Get(s string) string {
-	e.ml.Lock()
-	defer e.ml.Unlock()
-
-	if v, ok := e.mp[s]; !ok {
-		return ""
-	} else if v, ok := v.(string); !ok {
-		return ""
-	} else {
-		return v
-	}
-}
-
-// New returns a new Event with the options applied.
-func New(opts ...Option) *Event {
-	e := &Event{
-		mp: map[string]interface{}{
-			"date": time.Now(),
-		},
-	}
-
-	for _, opt := range opts {
-		opt(e)
-	}
-
-	return e
-}
 
 // Apply applies all options to the Event returning it after it's done.
 func Apply(e *Event, opts ...Option) *Event {
